+++ Drop-In/src/main/java/com/braintreepayments/api/DropInClient.java
@@ -7,7 +7,6 @@ import android.text.TextUtils;
 
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.FragmentActivity;
 
 import java.util.ArrayList;
@@ -20,10 +19,10 @@ import java.util.Set;
  */
 public class DropInClient {
 
-    public static final String EXTRA_CHECKOUT_REQUEST = "com.braintreepayments.api.EXTRA_CHECKOUT_REQUEST";
-    public static final String EXTRA_CHECKOUT_REQUEST_BUNDLE = "com.braintreepayments.api.EXTRA_CHECKOUT_REQUEST_BUNDLE";
-    public static final String EXTRA_SESSION_ID = "com.braintreepayments.api.EXTRA_SESSION_ID";
-    public static final String EXTRA_AUTHORIZATION = "com.braintreepayments.api.EXTRA_AUTHORIZATION";
+    static final String EXTRA_CHECKOUT_REQUEST = "com.braintreepayments.api.EXTRA_CHECKOUT_REQUEST";
+    static final String EXTRA_CHECKOUT_REQUEST_BUNDLE = "com.braintreepayments.api.EXTRA_CHECKOUT_REQUEST_BUNDLE";
+    static final String EXTRA_SESSION_ID = "com.braintreepayments.api.EXTRA_SESSION_ID";
+    static final String EXTRA_AUTHORIZATION = "com.braintreepayments.api.EXTRA_AUTHORIZATION";
 
     static final String LAST_USED_PAYMENT_METHOD_TYPE =
             "com.braintreepayments.api.dropin.LAST_USED_PAYMENT_METHOD_TYPE";
@@ -105,10 +104,10 @@ public class DropInClient {
                         callback.onResult(null, continueError);
                     } else if (threeDSecureResult != null) {
                         final DropInResult dropInResult = new DropInResult();
-                        dropInResult.paymentMethodNonce(threeDSecureResult.getTokenizedCard());
+                        dropInResult.setPaymentMethodNonce(threeDSecureResult.getTokenizedCard());
                         dataCollector.collectDeviceData(activity, (deviceData, dataCollectionError) -> {
                             if (deviceData != null) {
-                                dropInResult.deviceData(deviceData);
+                                dropInResult.setDeviceData(deviceData);
                                 callback.onResult(dropInResult, null);
                             } else {
                                 callback.onResult(null, dataCollectionError);
@@ -184,7 +183,7 @@ public class DropInClient {
         return braintreeClient.getBrowserSwitchResult(activity);
     }
 
-    public void deliverBrowserSwitchResult(final FragmentActivity activity, final DropInResultCallback callback) {
+    void deliverBrowserSwitchResult(final FragmentActivity activity, final DropInResultCallback callback) {
         BrowserSwitchResult browserSwitchResult = braintreeClient.deliverBrowserSwitchResult(activity);
         if (browserSwitchResult != null) {
             int requestCode = browserSwitchResult.getRequestCode();
@@ -246,8 +245,8 @@ public class DropInClient {
             return;
         }
 
-        final DropInResult dropInResult = new DropInResult()
-                .paymentMethodNonce(paymentMethodNonce);
+        final DropInResult dropInResult = new DropInResult();
+        dropInResult.setPaymentMethodNonce(paymentMethodNonce);
         dataCollector.collectDeviceData(activity, (deviceData, dataCollectionError) -> {
             if (dataCollectionError != null) {
                 callback.onResult(null, dataCollectionError);
@@ -255,7 +254,7 @@ public class DropInClient {
             }
 
             if (deviceData != null) {
-                dropInResult.deviceData(deviceData);
+                dropInResult.setDeviceData(deviceData);
                 callback.onResult(dropInResult, null);
             }
         });
@@ -338,6 +337,12 @@ public class DropInClient {
         });
     }
 
+    /**
+     * Called to launch a {@link DropInActivity}
+     *
+     * @param activity the current {@link FragmentActivity}
+     * @param requestCode the request code for the activity that will be launched
+     */
     public void launchDropInForResult(FragmentActivity activity, int requestCode) {
         Bundle dropInRequestBundle = new Bundle();
         dropInRequestBundle.putParcelable(EXTRA_CHECKOUT_REQUEST, dropInRequest);
@@ -355,7 +360,7 @@ public class DropInClient {
      * Note: a client token must be used and will only return a payment method if it contains a
      * customer id.
      *
-     * @param activity the current {@link AppCompatActivity}
+     * @param activity the current {@link FragmentActivity}
      * @param callback callback for handling result
      */
     public void fetchMostRecentPaymentMethod(FragmentActivity activity, final FetchMostRecentPaymentMethodCallback callback) {
@@ -391,7 +396,7 @@ public class DropInClient {
                 DropInResult result = new DropInResult();
                 if (paymentMethodNonceList.size() > 0) {
                     PaymentMethodNonce paymentMethod = paymentMethodNonceList.get(0);
-                    result.paymentMethodNonce(paymentMethod);
+                    result.setPaymentMethodNonce(paymentMethod);
                 }
                 callback.onResult(result, null);
             } else if (error != null) {