+++ Drop-In/src/main/java/com/braintreepayments/api/DropInResultCallback.java
@@ -6,7 +6,7 @@ import androidx.fragment.app.FragmentActivity;
 /**
  * Callback for receiving the result of {@link DropInClient#deliverBrowserSwitchResult(FragmentActivity, DropInResultCallback)}
  */
-public interface DropInResultCallback {
+interface DropInResultCallback {
     /**
      * @param dropInResult {@link DropInResult}
      * @param error an exception that occurred while processing a DropInResult